refactor(utility): created get_program_structure utility and moved out all dependencies from program routers
